Factor out `BufferedReader` in a standalone file
Move the `BufferedReader` class to a standalone file. This allows the
class to be reused in other packages.
